Location Tracking can now be enabled/disabled
