+++ src/components/Map.js
@@ -1,17 +1,9 @@
 import React from 'react';
 import {StyleSheet,Text} from 'react-native';
-import MapView, {Polyline} from 'react-native-maps';
+import MapView, {Polyline,Circle} from 'react-native-maps';
 import {connect} from 'react-redux';
 
 const Map =({LocationState})=>{
-  // let points =[];
-  //
-  // for (let i =0; i<20 ;i++){
-  //   points.push({
-  //     latitude: 37.33233 + i * 0.001,
-  //     longitude: -122.03121 +i *0.001
-  //   });
-  // };
   console.log(LocationState.currentLocation);
   return(
     <>
@@ -27,6 +19,12 @@ const Map =({LocationState})=>{
             latitudeDelta:0.01,
             longitudeDelta:0.01
           }}>
+          <Circle
+            center={LocationState.currentLocation.coords}
+            radius={30}
+            strokeColor="rgba(158,158,255,1.0)"
+            fillColor="rgba(158,158,255,0.3)"
+          />
         </MapView>
       : null
     }
