+++ src/screens/TrackCreateScreen.js
@@ -1,38 +1,15 @@
 import '../_mockLocation';
-import React, {useEffect, useState} from 'react';
+import React from 'react';
 import {View,StyleSheet, Button, TextInput,TouchableOpacity} from 'react-native';
 import {Text} from 'react-native-elements';
-import {SafeAreaView} from 'react-navigation';
-import {requestPermissionsAsync, watchPositionAsync, Accuracy} from  'expo-location';
+import {SafeAreaView, withNavigationFocus} from 'react-navigation';
 import Map from '../components/Map'
 import {connect} from 'react-redux';
 import {addLocation} from '../actions';
+import useLocation from '../hooks/useLocation';
 
-const TrackCreateScreen= ({addLocation})=>{
-  const [err,setErr]= useState(null);
-
-  const startWatching=async ()=>{
-    try{
-      await requestPermissionsAsync();
-      //watch users
-      await watchPositionAsync(
-        {
-          accuracy: Accuracy.BestForNavigation,
-          timeInterval:1000,
-          distanceInterval:10
-        },
-        (location) =>{
-            addLocation(location);
-        }
-    )}catch(e){
-      setErr(e);
-    }
-  };
-
-//runs once when component is first rendered
-  useEffect(()=>{
-    startWatching();
-  },[]);
+const TrackCreateScreen= ({isFocused, addLocation})=>{
+  const [err]= useLocation(isFocused, (location)=>addLocation(location));
 
   return (
     <SafeAreaView force Inset={{top: 'always'}}>
@@ -42,5 +19,6 @@ const TrackCreateScreen= ({addLocation})=>{
     </SafeAreaView>
 )};
 
+const createWrapped= withNavigationFocus(TrackCreateScreen);
 
-export default connect(null,{addLocation})(TrackCreateScreen);
+export default connect(null,{addLocation})(createWrapped);
